+++ src/components/Header.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import LinkWrapper from './LinkWrapper';
+import LinkWrapper from '../utils/LinkWrapper';
 
 const Header = () => {
     return (
